usesに複写・編集した各ファイルの内容を再複写
