--- app/helpers/uses_helper.rb
@@ -1,2 +1,0 @@
-module UsesHelper
-end
