+++ app/views/users/edit.html.erb
@@ -1,2 +1,6 @@
-<h1>Users#edit</h1>
-<p>Find me in app/views/users/edit.html.erb</p>
+<h3>タスク編集</h3>
+<br>
+<div>
+  <%= render 'form', task: @task %>
+  <br>
+</div>
