+++ app/views/users/new.html.erb
@@ -1,2 +1,6 @@
-<h1>Users#new</h1>
-<p>Find me in app/views/users/new.html.erb</p>
+<h3>ユーザー登録</h3>
+<br>
+<div>
+  <%= render 'form', user: @user %>
+  <br>
+</div>
