--- app/views/uses/edit.html.erb
@@ -1,6 +1,0 @@
-<h3>タスク編集</h3>
-<br>
-<div>
-  <%= render 'form', task: @task %>
-  <br>
-</div>
