--- app/views/uses/new.html.erb
@@ -1,6 +1,0 @@
-<h3>ユーザー登録</h3>
-<br>
-<div>
-  <%= render 'form', user: @user %>
-  <br>
-</div>
