--- app/views/uses/show.html.erb
@@ -1,2 +1,0 @@
-<h1>Uses#show</h1>
-<p>Find me in app/views/uses/show.html.erb</p>
